+++ test/analysis_options_test.dart
@@ -34,6 +34,26 @@ linter:
     - always_put_required_named_parameters_first
 ''';
 
+const _contentWitMetricsRulesAsMap = '''
+analyzer:
+  plugins:
+    - dart_code_metrics
+  strong-mode:
+    implicit-casts: false
+    implicit-dynamic: false
+
+dart_code_metrics:
+  rules:
+    double-literal-format: false
+    newline-before-return: true
+    no-boolean-literal-compare: false
+
+linter:
+  rules:
+    - always_put_control_body_on_new_line
+    - always_put_required_named_parameters_first
+''';
+
 void main() {
   group('AnalysisOptions from', () {
     test('empty content', () {
@@ -47,10 +67,17 @@ void main() {
       expect(options.rulesNames, isEmpty);
     });
 
-    test('content with metrics', () {
-      final options = AnalysisOptions.from(_contentWitMetrics);
+    group('content with metrics', () {
+      test('rules defined as list', () {
+        final options = AnalysisOptions.from(_contentWitMetrics);
 
-      expect(options.rulesNames, equals(['double-literal-format']));
+        expect(options.rulesNames, equals(['double-literal-format']));
+      });
+      test('rules defined as map', () {
+        final options = AnalysisOptions.from(_contentWitMetricsRulesAsMap);
+
+        expect(options.rulesNames, equals(['newline-before-return']));
+      });
     });
   });
 }