+++ test/rules/no_boolean_literal_compare_rule_test.dart
@@ -26,7 +26,11 @@ void main() {
 
   var f = exampleString?.isEmpty==true;
 
-  var g = exampleString.isEmpty == true;
+  var g = true == exampleString?.isEmpty;
+
+  var h = exampleString.isEmpty == true;
+
+  var i = true == exampleString.isEmpty;
 
   [true, false]
       .where((value) => value== false)
@@ -56,7 +60,7 @@ void main() {
     final issues = const NoBooleanLiteralCompareRule()
         .check(parseResult.unit, sourceUrl, parseResult.content);
 
-    expect(issues.length, equals(11));
+    expect(issues.length, equals(13));
 
     expect(issues.map((issue) => issue.ruleId).toSet().single,
         equals('no-boolean-literal-compare'));
@@ -64,14 +68,18 @@ void main() {
         equals(CodeIssueSeverity.style));
     expect(issues.map((issue) => issue.sourceSpan.sourceUrl).toSet().single,
         equals(sourceUrl));
-    expect(issues.map((issue) => issue.sourceSpan.start.offset),
-        equals([75, 96, 117, 137, 156, 176, 201, 242, 314, 353, 472]));
+    expect(
+        issues.map((issue) => issue.sourceSpan.start.offset),
+        equals(
+            [75, 96, 117, 137, 156, 176, 201, 242, 285, 327, 399, 438, 557]));
     expect(issues.map((issue) => issue.sourceSpan.start.line),
-        equals([7, 9, 11, 13, 15, 17, 19, 21, 24, 25, 34]));
+        equals([7, 9, 11, 13, 15, 17, 19, 21, 23, 25, 28, 29, 38]));
     expect(issues.map((issue) => issue.sourceSpan.start.column),
-        equals([11, 11, 11, 11, 7, 7, 11, 11, 25, 25, 34]));
-    expect(issues.map((issue) => issue.sourceSpan.end.offset),
-        equals([83, 104, 124, 147, 164, 185, 229, 271, 327, 366, 483]));
+        equals([11, 11, 11, 11, 7, 7, 11, 11, 11, 11, 25, 25, 34]));
+    expect(
+        issues.map((issue) => issue.sourceSpan.end.offset),
+        equals(
+            [83, 104, 124, 147, 164, 185, 229, 272, 314, 356, 412, 451, 568]));
     expect(
         issues.map((issue) => issue.sourceSpan.text),
         equals([
@@ -82,7 +90,9 @@ void main() {
           'e== true',
           'e !=false',
           'exampleString?.isEmpty==true',
+          'true == exampleString?.isEmpty',
           'exampleString.isEmpty == true',
+          'true == exampleString.isEmpty',
           'value== false',
           'value !=false',
           'value==true',
@@ -97,6 +107,8 @@ void main() {
           'Comparing boolean values to boolean literals is unnecessary, as those expressions will result in booleans too. Just use the boolean values directly or negate them.',
           'Comparing boolean values to boolean literals is unnecessary, as those expressions will result in booleans too. Just use the boolean values directly or negate them.',
           'Comparison of null-conditional boolean with boolean literal may result in comparing null with boolean.',
+          'Comparison of null-conditional boolean with boolean literal may result in comparing null with boolean.',
+          'Comparing boolean values to boolean literals is unnecessary, as those expressions will result in booleans too. Just use the boolean values directly or negate them.',
           'Comparing boolean values to boolean literals is unnecessary, as those expressions will result in booleans too. Just use the boolean values directly or negate them.',
           'Comparing boolean values to boolean literals is unnecessary, as those expressions will result in booleans too. Just use the boolean values directly or negate them.',
           'Comparing boolean values to boolean literals is unnecessary, as those expressions will result in booleans too. Just use the boolean values directly or negate them.',
@@ -112,6 +124,8 @@ void main() {
           'e',
           'e',
           'exampleString?.isEmpty ?? false',
+          'exampleString?.isEmpty ?? false',
+          'exampleString.isEmpty',
           'exampleString.isEmpty',
           '!value',
           'value',
@@ -127,6 +141,8 @@ void main() {
           'This expression is unnecessarily compared to a boolean. Just use it directly.',
           'This expression is unnecessarily compared to a boolean. Just use it directly.',
           'Prefer using null-coalescing operator with false literal on right hand side.',
+          'Prefer using null-coalescing operator with false literal on right hand side.',
+          'This expression is unnecessarily compared to a boolean. Just use it directly.',
           'This expression is unnecessarily compared to a boolean. Just use it directly.',
           'This expression is unnecessarily compared to a boolean. Just negate it.',
           'This expression is unnecessarily compared to a boolean. Just use it directly.',